requirements.txt
# Core dependencies for Agentic RAG with Autogen
pyautogen>=0.2.0
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
